remove appContainer
